+++ mode/guozhan/src/character/rest.js
@@ -39,46 +39,6 @@ export default {
 		dieAudios: ["pe_sunchen"],
 		hasSkinInGuozhan: false,
 	}),
-	gz_zhugedan: new Character({
-		sex: "male",
-		group: "ye",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_gongao"],
-	}),
-	gz_pot_weiyan: new Character({
-		sex: "male",
-		group: "ye",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_new_kuanggu"],
-	}),
-	gz_re_lvbu: new Character({
-		sex: "male",
-		group: "ye",
-		hp: 5,
-		maxHp: 5,
-		hujia: 0,
-		skills: ["gz_wuchang", "gz_liyu"],
-	}),
-	gz_yl_yuanshu: new Character({
-		sex: "male",
-		group: "ye",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gz_new_yongsi", "gz_new_weidi"],
-	}),
-	gz_jsrg_liuyan: new Character({
-		sex: "male",
-		group: "ye",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gz_tushe", "gz_limu"],
-	}),
 	// 分组2：谷爱凌
 	gz_tangzi: new Character({
 		sex: "male",
@@ -679,16 +639,6 @@ export default {
 		skills: ["gzbiaozhao", "gzyechou"],
 		doubleGroup: ["wu", "qun"],
 	}),
-
-	gz_yangwan: new Character({
-		sex: "female",
-		group: "shu",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gzyouyan", "gzzhuihuan"],
-		hasSkinInGuozhan: true,
-	}),
 	// 神谷爱凌
 	gz_key_ushio: new Character({
 		sex: "female",
@@ -733,231 +683,6 @@ export default {
 		hujia: 0,
 		skills: ["fakexunxi", "fakehuanjia"],
 		hasSkinInGuozhan: true,
-	}),
-	//国战无双
-	//阴的没边了十头人
-	gz_shichangshi: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 2,
-		maxHp: 2,
-		hujia: 0,
-		skills: ["gz_danggu", "gz_mowang"],
-		trashBin: ["sex:male_castrated"],
-		names: "张|让-赵|忠-孙|璋-毕|岚-夏|恽-韩|悝-栗|嵩-段|珪-郭|胜-高|望",
-	}),
-	gz_scs_zhangrang: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 1,
-		maxHp: 1,
-		hujia: 0,
-		isUnseen: true,
-		skills: ["gz_taoluan"],
-		trashBin: ["sex:male_castrated"],
-	}),
-	gz_scs_zhaozhong: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 1,
-		maxHp: 1,
-		hujia: 0,
-		isUnseen: true,
-		skills: ["gz_chiyan"],
-		trashBin: ["sex:male_castrated"],
-	}),
-	gz_scs_sunzhang: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 1,
-		maxHp: 1,
-		hujia: 0,
-		isUnseen: true,
-		skills: ["gz_zimou"],
-		trashBin: ["sex:male_castrated"],
-	}),
-	gz_scs_bilan: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 1,
-		maxHp: 1,
-		hujia: 0,
-		isUnseen: true,
-		skills: ["gz_picai"],
-		trashBin: ["sex:male_castrated"],
-	}),
-	gz_scs_xiayun: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 1,
-		maxHp: 1,
-		hujia: 0,
-		isUnseen: true,
-		skills: ["gz_yaozhuo"],
-		trashBin: ["sex:male_castrated"],
-	}),
-	gz_scs_hankui: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 1,
-		maxHp: 1,
-		hujia: 0,
-		isUnseen: true,
-		skills: ["gz_xiaolu"],
-		trashBin: ["sex:male_castrated"],
-	}),
-	gz_scs_lisong: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 1,
-		maxHp: 1,
-		hujia: 0,
-		isUnseen: true,
-		skills: ["gz_kuiji"],
-		trashBin: ["sex:male_castrated"],
-	}),
-	gz_scs_duangui: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 1,
-		maxHp: 1,
-		hujia: 0,
-		isUnseen: true,
-		skills: ["gz_chihe"],
-		trashBin: ["sex:male_castrated"],
-	}),
-	gz_scs_guosheng: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 1,
-		maxHp: 1,
-		hujia: 0,
-		isUnseen: true,
-		skills: ["gz_niqu"],
-		trashBin: ["sex:male_castrated"],
-	}),
-	gz_scs_gaowang: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 1,
-		maxHp: 1,
-		hujia: 0,
-		isUnseen: true,
-		skills: ["gz_miaoyu"],
-		trashBin: ["sex:male_castrated"],
-	}),
-	gz_gaolan: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_jungong", "gz_dengli"],
-		dieAudios: ["sp_gaolan"],
-	}),
-	gz_caoang: new Character({
-		sex: "male",
-		group: "wei",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["kaikang"],
-	}),
-	gz_zhangxingcai: new Character({
-		sex: "female",
-		group: "shu",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["shenxian", "gz_qiangwu"],
-	}),
-	gz_luzhi: new Character({
-		sex: "male",
-		group: "wei",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gz_qingzhong", "gz_weijing"],
-	}),
-	gz_quyi: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["fuqi", "jiaozi"],
-	}),
-	gz_caoying: new Character({
-		sex: "female",
-		group: "wei",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_lingren", "gz_fujian"],
-	}),
-	gz_guansuo: new Character({
-		sex: "male",
-		group: "shu",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_zhengnan"],
-	}),
-	gz_zhangxuan: new Character({
-		sex: "female",
-		group: "wu",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_tongli", "gz_shezang"],
-	}),
-	gz_zhangyao: new Character({
-		sex: "female",
-		group: "wu",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_yuanyu", "gz_xiyan"],
-	}),
-	gz_guanyinping: new Character({
-		sex: "female",
-		group: "shu",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["xueji", "gz_huxiao"],
-	}),
-	gz_xinxianying: new Character({
-		sex: "female",
-		group: "wei",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gz_caishi"],
-	}),
-	gz_yj_zhanghe: new Character({
-		sex: "male",
-		group: "qun",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_zhilve"],
-	}),
-	gz_mizhu: new Character({
-		sex: "male",
-		group: "shu",
-		hp: 3,
-		maxHp: 3,
-		hujia: 0,
-		skills: ["gz_ziyuan", "gz_jugu"],
-	}),
-	gz_caochun: new Character({
-		sex: "male",
-		group: "wei",
-		hp: 4,
-		maxHp: 4,
-		hujia: 0,
-		skills: ["gz_shanjia"],
 	}),
 	//OL限时武将
 	gz_re_xusheng: new Character({
