+++ mode/guozhan/src/skill/character/index.js
@@ -4,6 +4,7 @@ import shi from "./shi.js";
 import bian from "./bian.js";
 import quan from "./quan.js";
 import yingbian from "./yingbian.js";
+import wushuang from "./wushuang.js";
 
 import rest from "./rest.js";
 
@@ -15,4 +16,5 @@ export default {
 	...bian,
 	...quan,
 	...yingbian,
+	...wushuang,
 };
