+++ mode/guozhan/src/translate/character/index.js
@@ -2,10 +2,12 @@ import sort from "./sort.js";
 import normal from "./normal.js";
 import zhen from "./zhen.js";
 import yingbian from "./yingbian.js";
+import wushuang from "./wushuang.js";
 
 export default {
 	...sort,
 	...normal,
 	...zhen,
 	...yingbian,
+	...wushuang,
 };
