+++ mode/guozhan/src/translate/character/wushuang.js
@@ -1,0 +1,23 @@
+export default {
+	gz_db_wenyang: "文鸯",
+	gz_gaolan: "高览",
+	gz_caoang: "曹昂",
+	gz_zhangxingcai: "张星彩",
+	gz_luzhi: "鲁芝",
+	gz_quyi: "麹义",
+	gz_caoying: "曹婴",
+	gz_guansuo: "关索",
+	gz_zhangxuan: "张璇",
+	gz_guanyinping: "关银屏",
+	gz_xinxianying: "辛宪英",
+	gz_zhugedan: "诸葛诞",
+	gz_pot_weiyan: "魏延",
+	gz_mizhu: "糜竺",
+	gz_yl_yuanshu: "袁术",
+	gz_re_lvbu: "吕布",
+	gz_jsrg_liuyan: "刘焉",
+	gz_shichangshi: "十常侍",
+	gz_yj_zhanghe: "张郃",
+	gz_caochun: "曹纯",
+	gz_zhangyao: "张媱",
+};
