+++ mode/guozhan/src/translate/rest.js
@@ -553,119 +553,15 @@ export default {
 
 	gzchengshang: "承赏",
 	gzchengshang_info: "每回合限一次，当你使用指定了目标的牌结算完毕后，若你未因此牌造成过伤害，则你可以令其中一名目标角色交给你一张牌，若此牌和你使用的牌的花色和点数均相同，则你失去此技能。",
-	//官盗 国战无双
-	gz_db_wenyang: "文鸯",
-	gz_quedi: "却敌",
-	gz_quedi_info: "①你使用【杀】或【决斗】指定唯一目标时，可以选择一项：1.获得其一张手牌；2.弃置一张基本牌并令此牌造成伤害+1。②你杀死一名角色后，可以交换主副将。",
-	gz_zhuifeng: "椎锋",
-	gz_zhuifeng_info: "主将技，出牌阶段限一次，你可以视为使用一张【决斗】。",
-	gz_chongjian: "冲坚",
-	gz_chongjian_info: "副将技，你可以将一张装备牌当作无次数限制的【杀】或【酒】使用或打出。",
-	gz_gaolan: "高览",
-	gz_jungong: "峻攻",
-	gz_jungong_info: "出牌阶段限一次，你可以失去1点体力或弃置一张牌，视为使用一张无距离次数限制的【杀】。",
-	gz_dengli: "等力",
-	gz_dengli_info: "你使用【杀】指定其他角色为唯一目标，或成为其他角色使用【杀】的目标后，若你与其体力相等，你可以摸一张牌。",
-	gz_caoang: "曹昂",
-	gz_zhangxingcai: "张星彩",
-	gz_qiangwu: "枪舞",
-	gz_qiangwu_info: "出牌阶段限一次，你可以进行判定，本回合你使用的下一张点数大于判定结果的【杀】不计入次数，下一张点数小于判定结果的【杀】无距离限制。",
-	gz_luzhi: "鲁芝",
-	gz_qingzhong: "清忠",
-	gz_qingzhong_info: "出牌阶段开始时，你可以摸两张牌；此阶段结束时，你与手牌数最少的一名同势力或未确定势力的其他角色交换手牌。",
-	gz_weijing: "卫境",
-	gz_weijing_info: "每轮限一次，你可以将一张牌当【杀】或【闪】使用。",
-	gz_quyi: "麹义",
-	gz_caoying: "曹婴",
-	gz_lingren: "凌人",
-	gz_lingren_info: "每回合限一次，你使用伤害牌指定目标后，若其手牌数不大于你，你可选择一项：1.摸两张牌；2.令此牌对其造成的伤害+1。",
-	gz_fujian: "伏间",
-	gz_fujian_info: "准备阶段和结束阶段，你可以视为对一名手牌数不大于你的其他角色使用一张【知己知彼】。",
-	gz_guansuo: "关索",
-	gz_zhengnan: "征南",
-	gz_zhengnan_info: "其他角色死亡后，你可以摸三张牌并获得以下未拥有的一个技能：〖武圣〗〖当先〗〖制蛮〗。",
-	gz_zhangxuan: "张璇",
-	gz_tongli: "同礼",
-	gz_tongli_info: "每回合限一次，你使用基本或普通锦囊牌指定目标后，可以展示手牌。若展示牌颜色均相同，此牌额外结算一次。",
-	gz_shezang: "奢葬",
-	gz_shezang_info: "你首次进入濒死时，可以摸四张牌。",
-	gz_guanyinping: "关银屏",
-	gz_huxiao: "虎啸",
-	gz_huxiao_info: "锁定技，你造成火焰伤害后，令受伤角色摸一张牌，然后你本回合对其使用的下三张牌不计入次数。",
-	gz_xinxianying: "辛宪英",
-	gz_caishi: "才识",
-	gz_caishi_info: "摸牌阶段结束时，你可以选择一项：1.本回合手牌上限+2；2.回复1点体力，本回合不能对其他角色使用牌。",
-	gz_zhugedan: "诸葛诞",
-	gz_gongao: "功獒",
-	gz_gongao_info: "锁定技，其他角色死亡时，你增加1点体力上限并恢复1点体力；准备阶段，你将手牌摸至体力上限。",
-	gz_pot_weiyan: "魏延",
-	gz_new_kuanggu: "狂骨",
-	gz_new_kuanggu_info: "①当你造成或受到1点伤害后，你可以摸一张牌。②你使用单目标伤害牌可以额外指定至多X个目标（X为你已损失的体力值+1），",
-	gz_re_lvbu: "吕布",
-	gz_wuchang: "无常",
-	gz_wuchang_info: "锁定技，①你获得其他角色的牌后，本回合对该势力角色造成的伤害+1。②场上势力数减少后，你摸两张牌或将一张牌当作【出其不意】使用。",
-	gz_liyu: "利驭",
-	gz_liyu_info: "①出牌阶段开始时，你可以获得场上一张牌，然后失去牌的角色摸一张牌。②你使用【杀】的次数+X，你的【杀】需使用X张【闪】才能抵消（X为你装备区的牌数）。",
-	gz_yl_yuanshu: "袁术",
-	gz_new_yongsi: "庸肆",
-	gz_new_yongsi_info: "①摸牌阶段，你可以多摸X张牌（X为小势力存活角色数）。②出牌阶段限一次，你可以获得两名其他角色各一张牌，然后若你本回合未造成过伤害，你弃置两张牌并失去1点体力。",
-	gz_new_weidi: "伪帝",
-	gz_new_weidi_info: "锁定技，若你的宝物栏空置且场上没有【玉玺】，你视为装备了【玉玺】。",
-	gz_jsrg_liuyan: "刘焉",
-	gz_tushe: "图射",
-	gz_tushe_info: "你使用牌指定其他角色为目标时，可以摸X张牌（X为大势力角色数）。",
-	gz_limu: "立牧",
-	gz_limu_info: "出牌阶段限一次，你可以将一张♦牌当【乐不思蜀】对你使用，然后回复1点体力，令你本回合可以额外使用此牌点数张杀。",
-	gz_shichangshi: "十常侍",
-	gz_danggu: "党锢",
-	gz_danggu_info: "锁定技。①游戏开始时，你获得十张“常侍”牌，然后你亮出一张“常侍”牌。②当你修整结束后，你亮出一张“常侍”牌并摸一张牌。③若你有亮出的“常侍”牌，你视为拥有这些牌的技能。",
-	gz_mowang: "殁亡",
-	gz_mowang_info: "锁定技。①当你死亡前，若你有未亮出的“常侍”牌且体力上限大于0，你将死亡改为修整至你的下个回合开始前，然后你复原武将牌，且不于此次死亡事件中进行展示身份牌、检测游戏胜利条件与执行奖惩的流程。②回合结束后，你死亡。",
-	gz_taoluan: "滔乱",
-	gz_taoluan_info: "出牌阶段限一次。你可以将一张牌当任意一种基本牌或普通锦囊牌使用。",
-	gz_chiyan: "鸱咽",
-	gz_chiyan_info: "当你使用【杀】指定目标后，你可以依次令目标角色与你将任意张牌置于自己的武将牌上直到当前回合结束。若如此做，手牌数不大于/不小于你的目标角色本回合受到的伤害+1/不能使用手牌。",
-	gz_zimou: "自谋",
-	gz_zimou_info: "锁定技。出牌阶段开始时，你令所有角色选择一项：1.交给你一张牌；2.弃置你一张牌，然后受到你造成的1点伤害。",
-	gz_picai: "庀材",
-	gz_picai_info: "出牌阶段限一次。你可令X名角色依次将一张手牌置于牌堆顶，然后你亮出牌堆顶X张牌。其中每有一种类别，你便摸一张牌。若你因此获得三张牌，则因此失去牌的角色依次从亮出的牌中选择一张获得（X为你的体力值）。",
-	gz_yaozhuo: "谣诼",
-	gz_yaozhuo_info: "①出牌阶段限一次，你可以与一名角色拼点：若你赢，其弃置两张牌；若你没赢，你回复1点体力。②你拼点结算完成后，获得对方的拼点牌。",
-	gz_xiaolu: "宵赂",
-	gz_xiaolu_info: "出牌阶段限一次。你可以摸两张牌，然后选择一项：1.弃置两张手牌；2.将两张手牌交给一名其他角色。",
-	gz_kuiji: "窥机",
-	gz_kuiji_info: "出牌阶段限一次。你可以观看一名其他角色的手牌，然后弃置你与其的共计四张花色各不相同的手牌。",
-	gz_chihe: "叱吓",
-	gz_chihe_info: "当你使用【杀】指定唯一目标后，你可以摸两张牌并展示等量张手牌，然后你与其拼点，若你赢，此【杀】的伤害值基数+1；否则你弃置两张牌。",
-	gz_niqu: "逆取",
-	gz_niqu_info: "一名角色于你的出牌阶段使用【闪】后，你可以摸一张牌，然后视为对其使用一张【杀】。",
-	gz_miaoyu: "妙语",
-	gz_miaoyu_info: "你可以将♦牌当火【杀】使用；你使用【杀】无次数限制。",
-	gz_yj_zhanghe: "张郃",
-	gz_zhilve: "知略",
-	gz_zhilve_info: "出牌阶段限一次，你可以失去1点体力并令你本回合的手牌上限+1，然后选择一项：1.移动场上的一张牌；2.摸一张牌并视为使用一张无距离次数限制的【杀】。",
-	gz_mizhu: "糜竺",
-	gz_ziyuan: "资援",
-	gz_ziyuan_info: "出牌阶段限一次，你可以将任意张点数之和为13的手牌展示并交给一名其他同势力角色，然后该角色回复1点体力。",
-	gz_jugu: "巨贾",
-	gz_jugu_info: "锁定技，此武将牌明置时，你摸X张牌，手牌上限+X（X为你的体力上限）。",
-	gz_caochun: "曹纯",
-	gz_shanjia: "缮甲",
-	gz_shanjia_info: "出牌阶段开始时，你可以摸X张牌（X为与你同势力的角色数），然后弃置一张牌。若为装备牌，你视为使用一张【杀】。",
-	gz_zhangyao: "张媱",
-	gz_yuanyu: "怨语",
-	gz_yuanyu_info: "出牌阶段限一次，你可以摸一张牌并将一张手牌置于武将牌上，称为“怨”。",
-	gz_xiyan: "夕颜",
-	gz_xiyan_info: "你受到伤害后，若伤害牌和“怨”颜色相同，可以令伤害来源选择一项：1.本回合手牌上限-4；2.本回合不能使用基本牌。",
 
 	guozhan_jun: "君主武将",
 	guozhan_jin: "文德武备",
-	guozhan_jinEX1: "紫气东来",
-	guozhan_jinEX2: "受命于天",
+	guozhan_jinEX1: "紫气东来·受命于天",
+	//guozhan_jinEX2: "受命于天",
 	guozhan_single: "君临天下EX",
 	guozhan_double: "双势力武将",
 	guozhan_yexinjia: "野心家武将",
-	guozhan_zongheng: "纵横捭阖",
+	guozhan_zongheng: "十周年专属",//"纵横捭阖",
 	guozhan_online: "Online专属",
 	guozhan_tw: "海外服专属",
 	guozhan_mobile: "移动版专属",
